requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
supabase==2.0.0
python-jose==3.3.0
flask-jwt-extended==4.5.3
gunicorn==21.2.0
python-multipart==0.0.6
PyPDF2==3.0.1 